+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2017-04-09
+
+* Add control message support.
+* Add `IsClosed` to Session.
+
 ## 2017-02-10
 
 * Return errors for some exposed methods.